refactor: #49 카테고리 별도 도메인으로 분리
