+++ src/main/java/com/wholeseeds/mindle/domain/category/repository/custom/CategoryRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.wholeseeds.mindle.domain.category.repository.custom;
+
+import java.util.List;
+
+import com.wholeseeds.mindle.domain.category.dto.CategoryRow;
+
+public interface CategoryRepositoryCustom {
+	List<CategoryRow> findAllFlat();
+}
