--- src/main/java/com/wholeseeds/mindle/domain/complaint/repository/custom/CategoryRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.wholeseeds.mindle.domain.complaint.repository.custom;
-
-import java.util.List;
-
-import com.wholeseeds.mindle.domain.complaint.dto.CategoryRow;
-
-public interface CategoryRepositoryCustom {
-	List<CategoryRow> findAllFlat();
-}
